requirements.txt
PyQt6>=6.4.0
qrcode[pil]
Pillow
pyperclip
